requirements/base.txt
django==5.1.4
psycopg-binary==3.2.3
django-environ==0.11.2
django-unfold==0.49.1
pygments==2.19.1
pydantic==2.10.5
google-cloud-bigquery==3.28.0
google-cloud-bigquery-storage==2.28.0
py-avro-schema==3.8.2
django-ninja==1.3.0
djangorestframework==3.15.2
django-storages==1.14.4